requirements.txt
anyio
bcrypt
cachetools
SQLAlchemy~=2.0.7
certifi
charset-normalizer
click
colorama
ecdsa
fastapi~=0.95.0
google-api-core
google-api-python-client
google-auth==2.16.3
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
transliterate
typing_extensions
uritemplate
urllib3
uvicorn
wheel==0.40.0
gspread
h11
httplib2
idna
passlib~=1.7.4
pip
protobuf==4.22.1
requests
requests-oauthlib
setuptools==67.6.0
six
sniffio
starlette
psycopg2
PyYAML~=6.0